Handle version X.Y.Z-alphaW
