+++ its/org.sonarlint.eclipse.its/src/org/sonarlint/eclipse/its/NotificationsCheckboxTest.java
@@ -20,11 +20,6 @@
 package org.sonarlint.eclipse.its;
 
 import com.sonar.orchestrator.Orchestrator;
-import org.eclipse.swtbot.eclipse.finder.widgets.SWTBotView;
-import org.eclipse.swtbot.swt.finder.finders.UIThreadRunnable;
-import org.eclipse.swtbot.swt.finder.results.BoolResult;
-import org.eclipse.swtbot.swt.finder.waits.DefaultCondition;
-import org.eclipse.swtbot.swt.finder.widgets.SWTBotTreeItem;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Test;
@@ -98,23 +93,6 @@ public class NotificationsCheckboxTest extends AbstractSonarLintTest {
     assertThat(wizardBot.isNextEnabled()).isFalse();
     wizardBot.clickFinish();
 
-    SWTBotView serversView = bot.viewById("org.sonarlint.eclipse.ui.ServersView");
-    final SWTBotTreeItem serverCell = serversView.bot().tree().getAllItems()[0];
-    bot.waitUntil(new DefaultCondition() {
-      @Override
-      public boolean test() throws Exception {
-        return UIThreadRunnable.syncExec(new BoolResult() {
-          @Override
-          public Boolean run() {
-            return serverCell.getText().matches(connectionName + " \\[Version: " + orchestrator.getServer().version() + "(.*), Last storage update: (.*)\\]");
-          }
-        });
-      };
-
-      @Override
-      public String getFailureMessage() {
-        return "Server status is: " + serverCell.getText();
-      }
-    }, 20_000);
+    waitForServerUpdate(connectionName, orchestrator);
   }
 }
